+++ src/ApolloModule.ts
@@ -1,14 +1,8 @@
-import {
-  NgModule,
-  ModuleWithProviders,
-} from '@angular/core';
+import { NgModule, ModuleWithProviders } from '@angular/core';
+
+import { angularApolloClient, Angular2Apollo } from './Angular2Apollo';
 
 import ApolloClient from 'apollo-client';
-
-import {
-  angularApolloClient,
-  Angular2Apollo,
-} from './angular2Apollo';
 
 @NgModule({
   providers: [Angular2Apollo],
