+++ src/ApolloQueryPipe.ts
@@ -1,6 +1,4 @@
-import {
-  Pipe,
-} from '@angular/core';
+import { Pipe } from '@angular/core';
 
 @Pipe({
   name: 'apolloQuery',
