+++ tests/utils/observeVariables.ts
@@ -1,8 +1,4 @@
 import { Subject } from 'rxjs/Subject';
-import { Observable } from 'rxjs/Observable';
-
-import 'rxjs/add/observable/of';
-import 'rxjs/add/operator/switch';
 
 import { observeVariables } from '../../src/utils/observeVariables';
